Moved ToastManager outside of internal folder
