+++ src/managers/ToastManager/helpers/LifetimeManager.ts
