+++ src/managers/ToastManager/helpers/getToastDataStateByStatus.ts
