+++ src/managers/ToastManager/helpers/tests/getToastDataStateByStatus.test.ts
