+++ src/managers/ToastManager/tests/getToastProceededStatus.test.ts
