+++ src/managers/ToastManager/types/index.ts
