+++ src/managers/ToastManager/types/toast.types.ts
