+++ src/managers/internal/index.ts
@@ -1,4 +1,3 @@
 export { PendingTransactionsStateManager } from './PendingTransactionsStateManager/PendingTransactionsStateManager';
 export { SignTransactionsStateManager } from './SignTransactionsStateManager/SignTransactionsStateManager';
-export { ToastManager } from './ToastManager/ToastManager';
 export { WalletConnectStateManager as WalletConnectManager } from './WalletConnectStateManager/WalletConnectStateManager';
